End of 20210721

+++ API/DTOs/ProdDefinitionDto.cs
@@ -4,7 +4,7 @@ namespace API.DTOs
     {
         public int AutoId { get; set; } 
         public string PDName { get; set; } 
-        public int PDHeaderId { get; set; }           
+        public byte PDHeaderId { get; set; }           
         public int ProcessId { get; set; }  
         public int ReceiveSiteId { get; set; }        
         public int DispatchSiteId { get; set; }
