+++ API/Entities/MstrProductionDefinitionDt.cs
@@ -9,7 +9,7 @@ namespace API.Entities
     {
         [Key]
         public int AutoId { get; set; } 
-        public int PDHeaderId { get; set; }  
+        public byte PDHeaderId { get; set; }  
         public int ProcessId { get; set; }  
         public int ReceiveSiteId { get; set; }        
         public int DispatchSiteId { get; set; }
